+++ cubestring.h
@@ -89,9 +89,9 @@ public:
      *
      * @param _nextLetterDelta is the delta you want to apply from letter to letter angle. [default 0]
      *
-     * @return [GLint] the angle the next letter after this string should have using the :nextLetterDelta.
+     * @return [GLfloat] the angle the next letter after this string should have using the :nextLetterDelta.
      */
-    GLint setCurrentAngle(GLint _firstLetterAngle, GLint _nextLetterDelta = 0);
+    GLfloat setCurrentAngle(GLfloat _firstLetterAngle, GLfloat _nextLetterDelta = 0);
 
 
 private:
